+++ drydock/duster.py
@@ -62,6 +62,7 @@ class Container(object):
         """Return the docker command required to create this container."""
         cmd = ["docker run -d -dns 172.17.42.1"]
         cmd.append("-name " + self.name)
+        cmd.append("-h " + self.fqdn)
         cmd.extend(self.get_portmaps())
         if self.data:
             cmd.append("-v /mnt/drydock:/mnt/drydock")