+++ HISTORY.rst
@@ -6,6 +6,8 @@ History
 0.4.0 (TBD)
 ++++++++++++++++++
 
+* Added: ``data`` in subcontainer specification maps volumes at
+``/mnt/drydock``.
 * Fix: ``pull`` command also grabs the containers required for the
 ``prepare`` command.